adapted some code interface changes

+++ codeproject/How-to-extend-Strus-with-new-functions/src/minwin_summarizer.cpp
@@ -64,7 +64,7 @@ public:
 			const std::string& name_,
 			PostingIteratorInterface* postingIterator_,
 			const std::vector<SummarizationVariable>& /*variables_*/,
-			float /*weight_*/,
+			double /*weight_*/,
 			const TermStatistics& /*stats_*/)
 	{
 		try
