+++ codeproject/How-to-extend-Strus-with-new-functions/src/minwin_weighting.cpp
@@ -54,7 +54,7 @@ public:
 	virtual void addWeightingFeature(
 			const std::string& name_,
 			PostingIteratorInterface* postingIterator_,
-			float /*weight_*/,
+			double /*weight_*/,
 			const TermStatistics& /*stats_*/)
 	{
 		try
